Split god drop zone into distinct drop zones in vault list and vault detail unlocked
