+++ src/main/resources/fxml/vault_detail_unlocked.fxml
@@ -5,6 +5,7 @@
 <?import javafx.scene.control.Tooltip?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.Region?>
+<?import javafx.scene.layout.StackPane?>
 <?import javafx.scene.layout.VBox?>
 <VBox xmlns:fx="http://javafx.com/fxml"
 	  xmlns="http://javafx.com/javafx"
@@ -35,7 +36,18 @@
 
 	<Region VBox.vgrow="ALWAYS"/>
 
-	<HBox alignment="BOTTOM_RIGHT">
+	<HBox alignment="BOTTOM_CENTER">
+		<StackPane fx:id="dropZone">
+			<Button text="%main.vaultDetail.revealEncryptedItemBtn" minWidth="120" contentDisplay="BOTTOM">
+				<graphic>
+					<Label styleClass="label-small,label-muted" text="%main.vaultDetail.dropZone"/>
+				</graphic>
+			</Button>
+			<Region styleClass="drag-n-drop-indicator" visible="${controller.draggingUnlockedVaultContentOver}"/>
+		</StackPane>
+
+		<Region HBox.hgrow="ALWAYS"/>
+
 		<Button text="%main.vaultDetail.stats" minWidth="120" onAction="#showVaultStatistics" contentDisplay="BOTTOM">
 			<graphic>
 				<VBox spacing="6">
