+++ src/main/resources/i18n/strings.properties
@@ -330,10 +330,6 @@ main.minimizeBtn.tooltip=Minimize
 main.preferencesBtn.tooltip=Preferences
 main.debugModeEnabled.tooltip=Debug mode is enabled
 main.supporterCertificateMissing.tooltip=Please consider donating
-## Drag 'n' Drop
-main.dropZone.dropVault=Add this vault
-main.dropZone.dropUnlockedVaultContent=Reveal encrypted file or folder
-main.dropZone.unknownDragboardContent=If you want to add a vault, drag it to this window
 ## Vault List
 main.vaultlist.emptyList.onboardingInstruction=Click here to add a vault
 main.vaultlist.contextMenu.remove=Remove…
@@ -363,6 +359,8 @@ main.vaultDetail.throughput.idle=idle
 main.vaultDetail.throughput.kbps=%.1f kiB/s
 main.vaultDetail.throughput.mbps=%.1f MiB/s
 main.vaultDetail.stats=Vault Statistics
+main.vaultDetail.revealEncryptedItemBtn=Reveal Encrypted Item
+main.vaultDetail.dropZone=– Drop Zone –
 ### Missing
 main.vaultDetail.missing.info=Cryptomator could not find a vault at this path.
 main.vaultDetail.missing.recheck=Recheck
